+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changed
 -------
 
 - `mpservice.streamer.Streamer.transform`: parameter `concurrency` used to default to 1 (i.e. no concurrency), now defaults to higher numbers (i.e. with concurrency).
+- `mpservice.mpserver.{Sequential, Ensemble}` were renamed to `SequentialServlet` and `EnsembleServlet` respectively.
 
 Added or enhanced
 -----------------